+++ src/test/converter2/issues/gh1968.ts
@@ -1,0 +1,19 @@
+export class Foo {
+    /** getter */
+    get x() {
+        return 1;
+    }
+    /** setter */
+    set x(value: number) {
+        throw new Error();
+    }
+}
+
+export class Bar {
+    /** {@inheritDoc Foo.x} */
+    x = 1;
+    /** {@inheritDoc Foo.x:getter} */
+    y = 2;
+    /** {@inheritDoc Foo.x:setter} */
+    z = 3;
+}